Unique portfolio for each user

+++ PortfolioTrackerClient/Services/PortfolioService/PortfolioService.cs
@@ -1,5 +1,8 @@
-﻿using PortfolioTrackerShared.Models;
+﻿using Newtonsoft.Json;
+using PortfolioTrackerShared.Models;
 using System.Net.Http.Json;
+using System.Text;
+
 namespace PortfolioTrackerClient.Services.PortfolioService
 {
 	public class PortfolioService : IPortfolioService
@@ -18,34 +21,36 @@ namespace PortfolioTrackerClient.Services.PortfolioService
 
         #region Stock-CRUD
 
+        public User PortfolioOwner { get; set; } = new();
+
         public List<PortfolioStock> PortfolioStocks { get; set; }
 
         public event EventHandler<PortfolioChangedArgs>? PortfolioChanged;
 
-        public async Task InitializeAsync()
+        public async Task InitializePortfolioAsync(int userId)
         {
-            PortfolioStocks = await GetDatabaseStocks();
+            PortfolioStocks = await GetPortfolioStocks(userId);
         }
 
         public void OnPortfolioChanged(List<PortfolioStock> portfolioStocks, PortfolioStock? modifiedStock = null, PortfolioAction portfolioAction = 0)
         {
-            PortfolioChanged?.Invoke(this, new PortfolioChangedArgs(portfolioStocks, modifiedStock, portfolioAction));
+            PortfolioChanged?.Invoke(this, new PortfolioChangedArgs(portfolioStocks, modifiedStock));
         }
 
-        public async Task<PortfolioStock> AddStock(PortfolioStock stock)
+        public async Task<PortfolioStock> AddStock(PortfolioStock stock, int userId)
         {
-            var response = await _httpClient.PostAsJsonAsync($"{serverBaseDomain}/api/portfolio", stock);
+            var response = await _httpClient.PostAsJsonAsync($"{serverBaseDomain}/api/portfolio/add?userId={userId}", stock);
             var newStock = (await response.Content.ReadFromJsonAsync<ServiceResponse<PortfolioStock>>()).Data;
-            OnPortfolioChanged(PortfolioStocks = await GetDatabaseStocks(), PortfolioStocks.FirstOrDefault(s => s.Ticker == stock.Ticker), PortfolioAction.Added);
+            OnPortfolioChanged(PortfolioStocks = await GetPortfolioStocks(userId), PortfolioStocks.FirstOrDefault(s => s.Ticker == stock.Ticker), PortfolioAction.Added);
             return newStock;
         }
 
-        public async Task<bool> DeleteStock(string stockToDelete)
+        public async Task<bool> DeleteStock(string stockToDelete, int userId)
         {
-            var response = await _httpClient.DeleteAsync($"{serverBaseDomain}/api/portfolio/{stockToDelete}");
+            var response = await _httpClient.DeleteAsync($"{serverBaseDomain}/api/portfolio/delete/{stockToDelete}?userId={userId}");
 
             if (response.IsSuccessStatusCode)
-                OnPortfolioChanged(PortfolioStocks = await GetDatabaseStocks(), PortfolioStocks.FirstOrDefault(s => s.Ticker == stockToDelete), PortfolioAction.Deleted);
+                OnPortfolioChanged(PortfolioStocks = await GetPortfolioStocks(userId), null, PortfolioAction.Deleted);
 
             return response.IsSuccessStatusCode;
 
@@ -57,9 +62,9 @@ namespace PortfolioTrackerClient.Services.PortfolioService
             return response.Data;
         }
 
-        public async Task<List<PortfolioStock>> GetDatabaseStocks()
+        public async Task<List<PortfolioStock>> GetPortfolioStocks(int userId) 
         {
-            var response = await _httpClient.GetFromJsonAsync<ServiceResponse<List<PortfolioStock>>>($"{serverBaseDomain}/api/portfolio");
+            var response = await _httpClient.GetFromJsonAsync<ServiceResponse<List<PortfolioStock>>>($"{serverBaseDomain}/api/portfolio?userId={userId}");
             return response.Data;
         }
 
@@ -68,9 +73,9 @@ namespace PortfolioTrackerClient.Services.PortfolioService
         /// </summary>
         /// <param name="stock"></param>
         /// <returns>Update success</returns>
-        public async Task<ServiceResponse<PortfolioStock>> UpdateStock(PortfolioStock stock)
+        public async Task<ServiceResponse<PortfolioStock>> UpdateStock(PortfolioStock stock, int userId)
         {
-            var response = await _httpClient.PutAsJsonAsync($"{serverBaseDomain}/api/portfolio", stock);
+            var response = await _httpClient.PutAsJsonAsync($"{serverBaseDomain}/api/portfolio?userId={userId}", stock);
 
             if (response.IsSuccessStatusCode)
                 OnPortfolioChanged(PortfolioStocks, stock, PortfolioAction.Modified);
